+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#16](https://github.com/LaxarJS/ax-input-control/issues/16): ignore `title` attribute when showing validation message
 - [#10](https://github.com/LaxarJS/ax-input-control/issues/10): validation: fixed error style with label using for-attribute